check server for Qurat and Aanchal for student and teacher side, Aatika for admin side
